ppdCollectionListPPDs(): Add cast to int for size_t output of libcups3
cupsArrayGetCount() of libcups3 outputs size_t and not int, as
libcups2 did. Added a cast to int so that it works with both libcups
versions.

+++ ppd/ppd-collection.cxx
@@ -184,7 +184,7 @@ cups_array_t *				// O - List of PPD files
 ppdCollectionListPPDs(
 	cups_array_t  *ppd_collections, // I - Directories to search for PPDs
 					//     in
-	int    limit,		// I - Limit
+	int           limit,		// I - Limit
 	int           num_options,	// I - Number of options
 	cups_option_t *options,		// I - Options
 	cf_logfunc_t log,		// I - Log function
@@ -390,7 +390,7 @@ ppdCollectionListPPDs(
 		 "libppd: [PPD Collections] %s=\"%s\"", options[i].name,
 		 options[i].value);
 
-  if (limit <= 0 || limit > cupsArrayGetCount(ppdlist.PPDsByMakeModel))
+  if (limit <= 0 || limit > (int)cupsArrayGetCount(ppdlist.PPDsByMakeModel))
     count = cupsArrayGetCount(ppdlist.PPDsByMakeModel);
   else
     count = limit;
